gh-2022 Refactor/move convert POJO into BaseObject method
